+++ core/comps/contacts/drtv-contact-tag.js
@@ -2,9 +2,9 @@
 
 angular.module('cmContacts').directive('cmContactTag',[
     'cmUserModel',
-    '$rootScope', '$routeParams',
+    '$rootScope', '$routeParams', '$timeout',
     function (cmUserModel,
-              $rootScope, $routeParams){
+              $rootScope, $routeParams, $timeout){
         return {
             restrict: 'AE',
             scope: {
@@ -19,25 +19,10 @@ angular.module('cmContacts').directive('cmContactTag',[
                 }
             },
             controller: function($scope){
-                $scope.isTrusted = function(contact){
-                    return      contact.identity
-                            &&  cmUserModel.verifyTrust(contact.identity)
-                };
-                /**
-                 * edit contact
-                 * @param id
-                 */
-                $scope.editContact = function (contact) {
-                    if(contact.contactType != 'pending') {
-                        $rootScope.goTo('/contact/' + contact.id);
+                $scope.editContact = function () {
+                    if($scope.contact.contactType != 'pending') {
+                        $rootScope.goTo('/contact/' + $scope.contact.id);
                     }
-                };
-                /**
-                 * delete contact via model
-                 * @param id
-                 */
-                $scope.deleteContact = function (contact) {
-                    cmLogger.debug('deleteContact ' + contact.id);
                 };
             }
         }